fp maps return u64

+++ csf/src/fp/gomap/mod.rs
@@ -39,7 +39,7 @@ impl<GS: GroupSize, SS: SeedSize, S: BuildSeededHasher> GOMap<GS, SS, S> {
     /// 
     /// If the `key` was not in the input key-value collection given during construction,
     /// either [`None`] or a value assigned to other key is returned.
-    pub fn get_stats<K: Hash + ?Sized, A: stats::AccessStatsCollector>(&self, key: &K, access_stats: &mut A) -> Option<u8> {
+    pub fn get_stats<K: Hash + ?Sized, A: stats::AccessStatsCollector>(&self, key: &K, access_stats: &mut A) -> Option<u64> {
         let mut groups_before = 0u64;
         let mut level_nr = 0u32;
         loop {
@@ -51,7 +51,7 @@ impl<GS: GroupSize, SS: SeedSize, S: BuildSeededHasher> GOMap<GS, SS, S> {
             if self.array.content.get_bit(bit_index) {
                 access_stats.found_on_level(level_nr);
                 //return Some(unsafe{self.array.rank_unchecked(bit_index)} as u64);
-                return Some(self.values.get_fragment(self.array.rank(bit_index), self.bits_per_value) as u8);
+                return Some(self.values.get_fragment(self.array.rank(bit_index), self.bits_per_value));
             }
             groups_before += level_size_groups;
             level_nr += 1;
@@ -62,7 +62,7 @@ impl<GS: GroupSize, SS: SeedSize, S: BuildSeededHasher> GOMap<GS, SS, S> {
     /// 
     /// If the `key` was not in the input key-value collection given during construction,
     /// either [`None`] or a value assigned to other key is returned.
-    #[inline] pub fn get<K: Hash + ?Sized>(&self, key: &K) -> Option<u8> {
+    #[inline] pub fn get<K: Hash + ?Sized>(&self, key: &K) -> Option<u64> {
         self.get_stats(key, &mut ())
     }
 
@@ -70,7 +70,7 @@ impl<GS: GroupSize, SS: SeedSize, S: BuildSeededHasher> GOMap<GS, SS, S> {
     /// 
     /// If the `key` was not in the input key-value collection given during construction,
     /// it either panics or returns a value assigned to other key is returned.
-    #[inline] pub fn get_stats_or_panic<K: Hash + ?Sized, A: stats::AccessStatsCollector>(&self, key: &K, access_stats: &mut A) -> u8 {
+    #[inline] pub fn get_stats_or_panic<K: Hash + ?Sized, A: stats::AccessStatsCollector>(&self, key: &K, access_stats: &mut A) -> u64 {
         self.get_stats(key, access_stats).expect("Invalid access to a key outside the set given during construction.")
     }
 
@@ -78,7 +78,7 @@ impl<GS: GroupSize, SS: SeedSize, S: BuildSeededHasher> GOMap<GS, SS, S> {
     /// 
     /// If the `key` was not in the input key-value collection given during construction,
     /// it either panics or returns a value assigned to other key is returned.
-    #[inline] pub fn get_or_panic<K: Hash + ?Sized>(&self, key: &K) -> u8 {
+    #[inline] pub fn get_or_panic<K: Hash + ?Sized>(&self, key: &K) -> u64 {
         self.get_stats_or_panic(key, &mut ())
     }
